% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_env.R
\name{calc_env}
\alias{calc_env}
\title{Calculate input values for Stock Synthesis from a Dynamic Factor Analysis}
\usage{
calc_env(x, index, se, sigma_r = 1)
}
\arguments{
\item{x}{A data frame, often provided by Nick Tolimieri.}

\item{index}{A string specifying the column name for the variable of
interest.}

\item{se}{A string specifying the column name for the standard error.}

\item{sigma_r}{A real number specifying the variability around recruitment
deviations used in the Stock Synthesis model. This was 1.0 in 2019 and
2021 for sablefish even though the input value for the stock assessment
was 1.4 because Melissa Haltuch decided that the index was independent
from the input value of recruitment deviations used in the assessment and
that the values should not be corrected by it. Especially given the fact
that recruitment variability might be tuned and thus end up being a
different value than what was used to create the input data. Thus, the
default is 1.0, which leads to no correction.}
}
\value{
A data frame with three columns, year, obs, se_log.
}
\description{
Calculate input values for Stock Synthesis from a Dynamic Factor Analysis
}
\examples{
\dontrun{
calc_env(read.csv(
  fs::path(
    here::here(),
    "data-raw",
    "MARSS_SSH_DFA_DF1_Index_4Kelli-SablefishAssessment.csv"
  )
), "index", "SE") |> head()
calc_env(read.csv(
  fs::path(
    here::here(),
    "data-raw",
    "Dynamic_Factors_and_SE_1925-2020_VariMaxRotated.csv"
  )
), "DF1", "DF1SE") |> head()
calc_env(read.csv(
  fs::path(
    here::here(),
    "data-raw",
    "SSH_Index_Bayes_DF1_1925-2020.csv"
  )
) |> dplyr::mutate(year = 1925:2020),
  "DF1_mean",
  "u.SE"
) |> head()
calc_env(read.csv(
  fs::path(
    here::here(),
    "data-raw",
    "ROTATED_DYNAMIC_FACTORS_Q2_SSH_working_2019.csv"
  )
), "DF1", "DF1SE") |> head()
}
}
