% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_total_catch.R
\name{compare_model_gemm_catch}
\alias{compare_model_gemm_catch}
\title{Check model estimated total catch to the GEMM estimates}
\usage{
compare_model_gemm_catch(
  r4ss_output,
  common_name,
  dir = NULL,
  add_name = NULL,
  verbose = TRUE
)
}
\arguments{
\item{r4ss_output}{List of model objects created by \code{r4ss::SS_output()}.}

\item{common_name}{Common name of the species in the GEMM database.}

\item{dir}{Directory to save figure. If no directory is provided the
figure is not saved. If a \code{dir} is specified a figure of the model
estimated total mortality compared to the GEMM total mortality, a
figure of total mortality across all years for the model, a rda of
all model years landed and catch estimates by fleet, and a rda of the
model estimated catch compared to the GEMM from 2003+. Default is NULL.}

\item{add_name}{String that can be added to the saved figure names.}

\item{verbose}{A logical that specifies if you want to print messages and
warnings to the console. The default is \code{TRUE}.}
}
\description{
Check model estimated total catch to the GEMM estimates
}
\author{
Chantel Wetzel
}
