% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_total_catch.R
\name{compare_total_catch}
\alias{compare_total_catch}
\title{Check model estimated total catch to the GEMM estimates}
\usage{
compare_total_catch(
  dir = NULL,
  r4ss_output,
  years = NULL,
  common_name = "sablefish"
)
}
\arguments{
\item{dir}{Directory to save figure}

\item{r4ss_output}{List of model objects created by r4ss::SS_output()}

\item{years}{Vector of years to filter where years should be continous (2002:2023)}

\item{common_name}{Common name of the species in the GEMM database}
}
\description{
Check model estimated total catch to the GEMM estimates
}
