% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_comparison_table.R
\name{create_comparison_table}
\alias{create_comparison_table}
\title{Table of estimated param and derived quantities across models}
\usage{
create_comparison_table(dir, model_summary, model_names, add_name = NULL)
}
\arguments{
\item{dir}{Directory location to save the output rda file.}

\item{model_summary}{List object created by \code{r4ss::SSsummarize()}.}

\item{model_names}{Text string of model names to add to table columns}

\item{add_name}{Optional text string that can be appended to the saved rda
file.}
}
\description{
Table of estimated param and derived quantities across models
}
