% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model_objects.R
\name{create_model_tables}
\alias{create_model_tables}
\title{Create r4ss rda files required for the report document}
\usage{
create_model_tables(
  model_name,
  dir = here::here("model"),
  fleetnames = c("Trawl", "Hook-and-Line", "Pot", "Trawl Discards",
    "Hook-and-Line Discards", "Pot Discard", "Triennial", "AFSC Slope", "NWFSC Slope",
    "WCGBT", "Recruitment Index")
)
}
\arguments{
\item{model_name}{String model folder name.}

\item{dir}{Directory to save figures. The full directory is required and if
provided a folder will be created in this location where figures will be
saved.}

\item{fleetnames}{List of user specific fleet names to use in the figures or
tables.}
}
\description{
Create r4ss rda files required for the report document
}
\author{
Chantel Wetzel
}
