% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-model-output-.R
\name{create_rda}
\alias{create_rda}
\title{Create model rda}
\usage{
create_rda(base_model, dir)
}
\arguments{
\item{base_model}{Base model folder name}

\item{dir}{Directory where the base model is located}
}
\description{
Create model rda
}
