% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{Commercial catch data}
\format{
\subsection{\code{data_commercial_catch}}{

Catches as reported in PacFIN; so, actually landings not catches:
\describe{
\item{year}{4 integers specifying the year}
\item{AGENCY_CODE}{}
\item{PACFIN_GROUP_GEAR_CODE}{}
\item{catch_mt}{Catch in mt}
}
}

\subsection{\code{data_survey_catch}}{

Survey catches
}

\subsection{\code{data_survey_bio}}{

Survey biological samples
}
}
\description{
Commercial catch data
}
