% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-tv-wtatage-weighted.R
\name{estimate_tv_wtatage_weighted}
\alias{estimate_tv_wtatage_weighted}
\title{Estimate time-varying weight at age that is biomass weighted}
\usage{
estimate_tv_wtatage_weighted(
  data,
  dir = here::here(),
  max_age = 30,
  do_plots = FALSE
)
}
\arguments{
\item{data}{Dataframe of year, weights, sex, and ages.  The data object can be
created by \code{process_weight_at_age_survey} and/or \code{process_weight_at_age_fishery}.}

\item{dir}{Parent directory for the repository.}

\item{max_age}{An integer specifying the maximum age of the modeled data
in the Stock Synthesis model. All age data beyond this will be assigned
to the maximum age. This is typically the age beyond which data are sparse
and weight and length are essentially the same across ages.}

\item{do_plots}{TRUE/FALSE to create plots from the model}
}
\value{
A data frame in long format with time-varying weight-at-age data.
}
\description{
Estimate time-varying weight at age that is biomass weighted
}
\details{
Predict weight-at-age for fishery and survey data with an area covariate
where the model includes age, year, cohort, sex, and area. There is a
smoother on age, year is modeled as a random effect, cohort is also a random
effect, and sex is a linear predictor to estimate weight, i.e.,
(weight ~ 1 + s(age, by = sex) + (1|fcohort) + (1|fyear) + sex + area). The
model estimated weight-at-age parameters by sex, cohort, year, and area are
then biomass weighted via a prediction grid based on a sdmTMB index.
}
\author{
Kelli F. Johnson, Chantel Wetzel, and Eric Ward
}
