% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_information.R
\name{figure_information}
\alias{figure_information}
\title{Search for information on {r4ss} figures using grep}
\usage{
figure_information(regex, location = fs::path(model_location, "plots"))
}
\description{
Search through a directory "location" for figures and compile information
about each returned png file into a data frame. Information stored in the
data frame will be helpful for adding a figure to your document.
}
\seealso{
\itemize{
\item \code{\link[sa4ss:add_figure]{sa4ss::add_figure()}}
\item \code{\link[=cat]{cat()}} for printing information to the screen
}
}
\author{
Kelli F. Johnson
}
