% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_information.R
\name{file_to_text}
\alias{file_to_text}
\title{Guess the best beginning part of a caption given a file name}
\usage{
file_to_text(x)
}
\arguments{
\item{x}{A vector of strings. The strings can be a full file path or just
the base name of the files. The paths do not have to be valid, in fact
they do not have to be paths at all for the function to work.}
}
\value{
A vector of strings the same length as \code{x}. Often this vector is pasted
together with additional information to formulate captions for figures.
Note that the returned strings are basically a phrase with no trailing
space at the end of the phrase.
}
\description{
Take a file name for a png file, or just the base of the file name, made
using {r4ss} and guess what the beginning of the caption should be.
\code{\link[dplyr:case_when]{dplyr::case_when()}} is used in combination with \code{\link[=grepl]{grepl()}} to work through
the potential options.
}
\examples{
file_to_text("comp_lendat_flt")
file_to_text("comp_agedat_flt")
file_to_text(c("comp_agedat_flt", "comp_agedat_flt"))
}
\seealso{
\itemize{
\item \code{\link[=figure_information]{figure_information()}}
}
}
\author{
Kelli F. Johnson
}
