% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_biomass_model.R
\name{fit_biomass_model}
\alias{fit_biomass_model}
\title{Fit sdmTMB model to WCGBT data to be used to estimate biomass weighted
weight-at-age estimates by are.}
\usage{
fit_biomass_model()
}
\description{
Fit sdmTMB model to WCGBT data to be used to estimate biomass weighted
weight-at-age estimates by are.
}
\author{
Eric Ward and Chantel Wetzel
}
