% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_values.R
\name{get_values}
\alias{get_values}
\title{Function to pull values from the read in report file and calculate the}
\usage{
get_values(replist, label, yrs, ci_value, single = FALSE)
}
\arguments{
\item{replist}{r4ss model output list created by r4ss::SS_output()}

\item{label}{description}

\item{yrs}{Vector of years.}

\item{ci_value}{To calculate confidence intervals, the desired interval must
be specified. The default is 0.95.}

\item{single}{Default is FALSE}
}
\value{
Individual .csv files for each executive summary table and additional tables
(catch, timeseries, numbers-at-age).
}
\description{
Function to pull values from the read in report file and calculate the
}
\author{
Chantel R. Wetzel, Kelli F. Johnson, Ian G. Taylor
}
