% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-.R
\name{load_data}
\alias{load_data}
\alias{load_data_catch}
\alias{load_data_catch_commercial}
\alias{load_data_catch_survey}
\title{Load data saved in \code{data-raw}}
\usage{
load_data_catch(directory = here::here())

load_data_catch_commercial(directory = here::here())

load_data_catch_survey(directory = here::here())
}
\arguments{
\item{directory}{The path to the assessment package of interest. The default
uses \code{\link[here:here]{here::here()}} to locate the appropriate directory but you can set it
to whatever directory you want as long as that directory contains
\code{data-raw}.}
}
\value{
All \verb{load_data_*()} functions return a data frame.
}
\description{
It is assumed that only the best files will be saved in \code{data-raw} such that
searching for something like \code{"PacFIN_SABL"} will only come up with two
files, one containing catch and the other containing biological data. So,
old outdated data should \strong{NOT} be saved in \code{data-raw}.
}
\examples{
\dontrun{
data_catch_survey <- load_data_catch_survey()
data_catch_commercial <- load_data_catch_commercial()
}
}
