% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_end_year_selex.R
\name{plot_end_year_selex}
\alias{plot_end_year_selex}
\title{Plot final year age-based selectivity, retention, and discard
mortality}
\usage{
plot_end_year_selex(replist, fleets = 1:3)
}
\arguments{
\item{replist}{List created by \code{r4ss::SS_output()}}

\item{fleets}{Vector of fleets to plot.}
}
\description{
Plot final year age-based selectivity, retention, and discard
mortality
}
