% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ageselex.R
\name{plot_fleet_retention}
\alias{plot_fleet_retention}
\title{Plot Length-Based Retention Curve}
\usage{
plot_fleet_retention(model_out, model_path, fleet_num)
}
\arguments{
\item{model_out}{model object created by \code{\link[r4ss:SS_output]{r4ss::SS_output()}}}

\item{model_path}{path to model directory containing control.ss file}

\item{fleet_num}{the fleet for which to plot selectivity}
}
\value{
ggplot plot object
}
\description{
Create a plot of retention-at-age for a given fishery or survey
fleet. Plots show selectivity curves for all model timeblocks.
}
\examples{

fleet_num <- 1 
mod_out <- r4ss::SS_output(dir=model_dir)
plot_fleet_retention(model_out, model_dir, fleet_num)

ret1 <- plot_fleet_retention(model_out, 1)
ret2 <- plot_fleet_retention(model_out, 2)
ret3 <- plot_fleet_retention(model_out, 3)
(ret1+ret2+ret3)+plot_layout(axes="collect")

}
