% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ageselex.R
\name{plot_fleet_selectivity}
\alias{plot_fleet_selectivity}
\title{Plot Age-Based Selectivity Curve}
\usage{
plot_fleet_selectivity(model_out, fleet_num)
}
\arguments{
\item{model_out}{model object created by \code{\link[r4ss:SS_output]{r4ss::SS_output()}}}

\item{fleet_num}{the fleet for which to plot selectivity}

\item{model_path}{path to model directory containing control.ss file}
}
\value{
ggplot plot object
}
\description{
Create a plot of selectivity-at-age for a given fishery or survey
fleet. Plots show selectivity curves for all model timeblocks and
available sexes.
}
\examples{

fleet_num <- 1 
mod_out <- r4ss::SS_output(dir=model_dir)
plot_fleet_selectivity(model_out, model_dir, fleet_num)


sel1 <- plot_fleet_selectivity(model_out, 1)
sel2 <- plot_fleet_selectivity(model_out, 2)
sel3 <- plot_fleet_selectivity(model_out, 3)
sel4 <- plot_fleet_selectivity(model_out, 4)
sel5 <- plot_fleet_selectivity(model_out, 5)
sel6 <- plot_fleet_selectivity(model_out, 6)+theme(legend.position.inside=c(0.65, 0.3))
sel7 <- plot_fleet_selectivity(model_out, 7)
sel7_leg <- sel7+guides(shape=guide_legend(), linetype=guide_legend())
legend <- cowplot::get_plot_component(sel7_leg, 'guide-box-inside', return_all = TRUE)
(sel1+sel2+sel3+sel4+sel5+sel6+sel7+legend$grob[[2]])+plot_layout(axes="collect")

}
