% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_comp_fits.R
\name{plot_ghostfleets}
\alias{plot_ghostfleets}
\title{Create sex-specific implied fits to length and age data in ghost fleets (e.g., negative fleet number)}
\usage{
plot_ghostfleets(replist, dir = NULL)
}
\arguments{
\item{replist}{List of model objects created by \code{r4ss::SS_output()}.}

\item{dir}{Directory to save figures. The full directory is required and if
provided a folder will be created in this location where figures will be
saved.  The default is NULL which will create a folder to save special
figures inside the replist directory.}

\item{common_name}{Common name of the species in the GEMM database.}
}
\description{
Create sex-specific implied fits to length and age data in ghost fleets (e.g., negative fleet number)
}
\author{
Chantel Wetzel
}
