% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_historical_ssb.R
\name{plot_historical_depl}
\alias{plot_historical_depl}
\title{Plot historical stock status}
\usage{
plot_historical_depl(historical_ssb)
}
\arguments{
\item{historical_ssb}{Dataframe created by data_historical-SSB.R}
}
\description{
Plot historical stock status
}
