% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_historical_ssb.R
\name{plot_historical_ssb}
\alias{plot_historical_ssb}
\title{Plot historical spawning outptu}
\usage{
plot_historical_ssb(historical_ssb)
}
\arguments{
\item{historical_ssb}{Dataframe created by data_historical-SSB.R}
}
\description{
Plot historical spawning outptu
}
