% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_comp_fits.R
\name{plot_len_fits_sexed_only}
\alias{plot_len_fits_sexed_only}
\title{Create sexed aggregated fits to length data (e.g., do not include unsexed data)}
\usage{
plot_len_fits_sexed_only(replist, dir = NULL, years = NULL)
}
\arguments{
\item{replist}{List of model objects created by \code{r4ss::SS_output()}.}

\item{dir}{Directory to save figures. The full directory is required and if
provided a folder will be created in this location where figures will be
saved.  The default is NULL which will create a folder to save special
figures inside the replist directory.}

\item{years}{A vector of continuous years to plot the aggregated fits for. Providing
a vector of years to plot can be useful to investigate fits for a subset of years
(e.g., block periods). The default is NULL which will plot the aggregated fleets for all years.}
}
\description{
Create sexed aggregated fits to length data (e.g., do not include unsexed data)
}
\author{
Chantel Wetzel
}
