% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ageselex.R
\name{plot_sel_ret_discard}
\alias{plot_sel_ret_discard}
\title{Plot Selectivity, Retention and Discard Mortality}
\usage{
plot_sel_ret_discard(model_out, model_path, sexes = 1:2, year = 2024)
}
\arguments{
\item{model_out}{model object created by \code{\link[r4ss:SS_output]{r4ss::SS_output()}}}

\item{model_path}{path to model directory containing control.ss file}

\item{sexes}{sexes across which plot (default 1:2)}

\item{years}{year to plot data for (default 2024)}
}
\value{
ggplot plot object
}
\description{
Create a plot of selectivity, retention, discard mortality
-at-age across sexes and fishing fleets
}
\examples{

mod_out <- r4ss::SS_output(dir=model_dir)

plot_sel_ret_discard(mod_out, model_dir, sexes=1:2, year=2024)

}
