% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_weight_at_age-.R
\name{plot_weight_age}
\alias{plot_weight_age}
\title{Plot weight-at-age data}
\usage{
plot_weight_age(data)
}
\arguments{
\item{data}{A data frame of weight-at-age values.}
}
\value{
A {ggplot2} object.
}
\description{
Plot weight-at-age data
}
\author{
Chantel Wetzel and Kelli F. Johnson
}
