% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_end_year_selex.R
\name{plot_year_selex}
\alias{plot_year_selex}
\title{Plot final year age-based selectivity, retention, and discard
mortality for a selected year.}
\usage{
plot_year_selex(replist, fleets = 1:3, year = 2024)
}
\arguments{
\item{replist}{List created by \code{r4ss::SS_output()}}

\item{fleets}{Vector of fleets to plot.}

\item{year}{Single numeric year to plot the selectivity for.}
}
\description{
Plot final year age-based selectivity, retention, and discard
mortality for a selected year.
}
