% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_bds.R
\name{process_bds_data}
\alias{process_bds_data}
\title{Process PacFIN bds data}
\usage{
process_bds_data(
  bds_data,
  catch_file = "data_commercial_catch_expansions.csv",
  age_bins = age_bins,
  length_bins = len_bins,
  gears = gears,
  common_name = common_name
)
}
\arguments{
\item{bds_data}{Data file of th PacFIN bds data.}

\item{catch_file}{Name of the catch file to be used to expand biological data.
This file should be located within \code{data-processed} with the \code{here} package
used to find the file in the directory.}

\item{age_bins}{Numerical sequence of age bins to process age data.}

\item{length_bins}{Numerical sequence of length bins to process length data.}

\item{gears}{Gear types to group the data within.  This should align with
PacFIN gear names.}

\item{common_name}{Argument input for pacfintools::cleanPacFIN() that defines
the gear-area fleet structure (coastwide HKL, Pot, and TWL).}
}
\description{
Process PacFIN bds data
}
