% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_survey.R
\name{process_survey}
\alias{process_survey}
\title{Process survey data}
\usage{
process_survey(
  catch_data,
  bds_data,
  strata,
  age_bins = age_bins,
  length_bins = len_bins,
  save_dir = "data-processed",
  fig_table_dir = here::here("data-raw", "survey", "trawl")
)
}
\arguments{
\item{catch_data}{Survey catch data frame}

\item{bds_data}{Survey bds data fram}

\item{strata}{Survey specific stratification}

\item{age_bins}{Age bins}

\item{length_bins}{Length bins}

\item{save_dir}{Directory to save processed data for the model. Default is
"data-processed"}

\item{fig_table_dir}{Directory to save figures and diagnostic tables.
Default is \code{here::here("data-raw", "survey", "trawl")}}
}
\description{
Process raw survey data into useable objects and save tables and figures
to data-processed, assessment/52tables, and assessment/53figures.
}
\section{Steps}{
\enumerate{
\item Create a table with number and proportion of positive tows, number of
lengths, and number of ages by project and year.
\item Calculate the design-based indices.
\item Calculate marginal length- and age-composition data for sexed and unsexed
fish. For 2021 and previous assessments, the unsexed fish were assigned
to sex using the sex ratio calculated in the second stage of the
expansion. The code now calculates marginals for sexed and then unsexed
fish as two separate data frames. The marginal ages for the WCGBTS are
not actually used in the model.
\item Calculate the conditional-composition data, where these ages are
conditional on a length bin.
}
}

