% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_weight_at_age-.R
\name{process_weight_at_age_fishery}
\alias{process_weight_at_age_fishery}
\title{Get the weight-at-age data from the fishery data.}
\usage{
process_weight_at_age_fishery(savedir = getwd())
}
\arguments{
\item{savedir}{A string specifying the path of interest.}

\item{bds}{}
}
\description{
Get the weight-at-age data from the fishery data.
}
\author{
Chantel Wetzel
}
