% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode-.R
\name{recode_project}
\alias{recode_project}
\title{Recode utility functions to convert from character string to fleet numbers.}
\usage{
recode_project(x, gls = FALSE)
}
\arguments{
\item{x}{String}

\item{gls}{Option to glue the recoded object. Default of FALSE}
}
\description{
Recode utility functions to convert from character string to fleet numbers.
}
