% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_data_sensitivities.R
\name{run_data_sensitivies}
\alias{run_data_sensitivies}
\title{Remove data sensitivies}
\usage{
run_data_sensitivies(
  dir,
  base_model,
  sens_dir,
  fleet,
  run_name,
  data_file,
  ctl_file,
  verbose = TRUE
)
}
\arguments{
\item{dir}{The directory where the base_model should be located}

\item{base_model}{The base name that sensitivities will be run for.}

\item{sens_dir}{The directory to save sensitivity runs}

\item{fleet}{The fleet number that the index, length, and age data will be
removed for.}

\item{run_name}{The run folder name.}

\item{data_file}{The name of the data file.}

\item{ctl_file}{The name of the control file.}

\item{verbose}{TRUE/FALSE to print messages to the screen.}
}
\description{
Remove data sensitivies
}
