% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sensitivites.R
\name{run_sensitivities}
\alias{run_sensitivities}
\title{Run the following sensitivities:}
\usage{
run_sensitivities(
  dir,
  base_model,
  sens_dir = NULL,
  data_file = "data.ss",
  ctl_file = "control.ss",
  verbose = TRUE
)
}
\arguments{
\item{dir}{The directory where the base_model should be located}

\item{base_model}{The base name that sensitivities will be run for.}

\item{sens_dir}{The directory to save sensitivity runs}

\item{data_file}{The name of the data file.}

\item{ctl_file}{The name of the control file.}

\item{verbose}{TRUE/FALSE to print messages to the screen.}
}
\description{
Run the following sensitivities:
}
