% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sablefish-package.R
\docType{package}
\name{sablefish}
\alias{sablefish}
\title{sablefish}
\description{
The assessment of the status of sablefish off of the U.S. West Coast hinges
on information provided from the results of analyses conducted within this
package. Functions are available to source the data, summarize the data, and
fit the data to Stock Synthesis models.
}
