% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sens_mirror_m.R
\name{sens_mirror_m}
\alias{sens_mirror_m}
\title{Mirror M between sexes sensitivies}
\usage{
sens_mirror_m(dir, base_model, sens_dir, run_name, ctl_file, verbose = TRUE)
}
\arguments{
\item{dir}{The directory where the base_model should be located}

\item{base_model}{The base name that sensitivities will be run for.}

\item{sens_dir}{The directory to save sensitivity runs}

\item{run_name}{The run folder name.}

\item{ctl_file}{The name of the control file.}

\item{verbose}{TRUE/FALSE to print messages to the screen.}
}
\description{
Mirror M between sexes sensitivies
}
