% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_composition_weights.R
\name{table_composition_weights}
\alias{table_composition_weights}
\title{Table of parameters}
\usage{
table_composition_weights(
  model,
  caption =
    "Data weightings applied to length and age compositions according to the `Francis' method.",
  label = "composition-weight"
)
}
\arguments{
\item{model}{A list from \code{\link[r4ss:SS_output]{r4ss::SS_output()}}.}

\item{caption}{A character string for the caption.}

\item{label}{A character string that will be used to specify the label.
Underscores are not allowed.}
}
\description{
Table of parameters
}
