% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_decision.R
\name{table_decision}
\alias{table_decision}
\title{Format a decision table}
\usage{
table_decision(
  ...,
  years = 2025:2036,
  format = c("latex", "html"),
  caption = formals(kableExtra::kbl)$caption,
  label = formals(kableExtra::kbl)$label,
  digits = c(0, 2, 3),
  tex = TRUE
)
}
\arguments{
\item{...}{Each element provided in the call to this function that is not
assigned to a particular input argument should be a vector of file paths
to the models you want to read for a given column of your decision table.
For example, if a single column has results from three models and
if your decision table has a low and high state of nature, then
there should be two vectors passed via \code{...} and each vector would contain
paths to three different models.
The order of the vectors will be the column order from left to right.
The order of the models within a vector will be the row order.}

\item{years}{A vector of years you want catches for.}

\item{format}{Output type.}

\item{caption}{A character string that
will be passed to the \code{caption} parameter of \link[kableExtra:kbl]{kableExtra::kbl}.
The default value is \code{NULL}.}

\item{label}{A character string without underscores that
will be passed to the \code{label} parameter of \link[kableExtra:kbl]{kableExtra::kbl}.
The default value is \code{NULL}.}

\item{digits}{A vector of digits for catch, spawning output, and fraction,
unfished. It gets passed to \code{r4ss::SS_decision_table_stuff()}.}

\item{tex}{TRUE/FALSE controlling whether to apply latex formatting.}
}
\description{
Format a decision table for its inclusion in a document.
}
\examples{
table_decision(
  list(mod.low.A, mod.base.A, mod.high.A),
  list(mod.low.B, mod.base.B, mod.high.B),
  list(mod.low.C, mod.base.C, mod.high.C)
)
}
\author{
Kelli F. Johnson, Ian G. Taylor, Chantel R. Wetzel
}
