% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_likelihoods.R
\name{table_likelihoods}
\alias{table_likelihoods}
\title{Table of likelihoods by source}
\usage{
table_likelihoods(
  model,
  caption = "Likelihood components by source.",
  label = "likelihoods"
)
}
\arguments{
\item{model}{A list from \code{\link[r4ss:SS_output]{r4ss::SS_output()}}.}

\item{caption}{A character string for the caption.}

\item{label}{A character string that will be used to specify the label.
Underscores are not allowed.}
}
\value{
A {kableExtra} table with two columns, Label and Total.
}
\description{
A {kableExtra} table for inclusion in a document of the likelihoods for each
data source in your model.
}
\author{
Ian G. Taylor
}
