% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_at_age-.R
\name{weight_at_age_outlier}
\alias{weight_at_age_outlier}
\title{Identify weight-at-age outliers}
\usage{
weight_at_age_outlier(data, filter = TRUE, drop = TRUE)
}
\arguments{
\item{data}{A data frame of weight-at-age values.}

\item{filter}{A logical specifying if the outliers should be removed or not.
The default is to remove them.}

\item{drop}{A logical specifying if the new column called \code{outlier} should
be removed from the data frame prior to returning it.}
}
\value{
A data frame with the same columns or one more column, i.e.,
\code{outlier}, than the input data frame.
}
\description{
Get weight-at-age data and identify any outliers using various screening
methods.
}
\seealso{
Other weight-at-age: 
\code{\link{weight_at_age_read}()}
}
\author{
Chantel Wetzel and Kelli F. Johnson
}
\concept{weight-at-age}
