% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_comparison_table.R
\name{create_comparison_table}
\alias{create_comparison_table}
\title{Table of estimated param and derived quantities across models}
\usage{
create_comparison_table(dir, model_list, model_names, add_name = NULL)
}
\arguments{
\item{dir}{Directory}

\item{model_list}{List of model folder names.  These folders should be inside
the directory.}

\item{model_names}{Text string of model names to add to figure legend.}

\item{add_name}{Optional text string that can be appended to the saved csv
file.}
}
\description{
Table of estimated param and derived quantities across models
}
