% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{Commercial landings data}
\format{
\subsection{\code{data_commercial_catch}}{

Landings for all model years:
\describe{
\item{year}{4 integers specifying the year}
\item{state}{landing state}
\item{area} {"north" and "south" of 36 degree N. lat.}
\item{gear_group} {gear type hkl, pot, trawl}
\item{catch_mt}{Landings in mt}
}
}

\subsection{\code{data_survey_catch}}{

Survey catches
}

\subsection{\code{data_survey_bio}}{

Survey biological samples
}
}
\description{
Commercial landings data
}
