% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-tv-wtatage-weighted.R
\name{estimate_tv_wtatage_weighted}
\alias{estimate_tv_wtatage_weighted}
\title{Estimate time-varying weight at age that is biomass weighted}
\usage{
estimate_tv_wtatage_weighted(
  dir = here::here(),
  max_age = 30,
  data_file = "data_weight_at_age_survey",
  do_plots = FALSE
)
}
\arguments{
\item{dir}{Parent directory for the repository.}

\item{max_age}{An integer specifying the maximum age of the modeled data
in the Stock Synthesis model. All age data beyond this will be assigned
to the maximum age. This is typically the age beyond which data are sparse
and weight and length are essentially the same across ages.}

\item{data_file}{Character string name of file with the weight at age data
to use. This file is expected to be a csv file stored in the "data-processed"
folder. Default is "data_weight_at_age_survey", but the option of
c("data_weight_at_age_survey", "data_weight_at_age_fishery") will include
both fishery and survey data.}

\item{do_plots}{TRUE/FALSE to create plots from the model}
}
\value{
A data frame in long format with time-varying weight-at-age data.
}
\description{
Estimate time-varying weight at age that is biomass weighted
}
\details{
Predict weight-at-age for fishery and survey data with an area covariate
where the model includes age, year, cohort, sex, and area. There is a
smoother on age, year is modeled as a random effect, cohort is also a random
effect, and sex is a linear predictor to estimate weight, i.e.,
(weight ~ 1 + s(age, by = sex) + (1|fcohort) + (1|fyear) + sex + area). The
model estimated weight-at-age parameters by sex, cohort, year, and area are
then biomass weighted via a prediction grid based on a sdmTMB index.
}
\author{
Kelli F. Johnson, Chantel Wetzel, and Eric Ward
}
