% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_survey.R
\name{process_survey}
\alias{process_survey}
\title{Process survey data}
\usage{
process_survey()
}
\description{
Process raw survey data into useable objects and save tables and figures
to data-processed, assessment/52tables, and assessment/53figures.
}
\section{Steps}{
\enumerate{
\item Create a table with number and proportion of positive tows, number of
lengths, and number of ages by project and year.
\item Calculate the design-based indices.
\item Calculate marginal length- and age-composition data for sexed and unsexed
fish. For 2021 and previous assessments, the unsexed fish were assigned
to sex using the sex ratio calculated in the second stage of the
expansion. The code now calculates marginals for sexed and then unsexed
fish as two separate data frames. The marginal ages for the WCGBTS are
not actually used in the model.
\item Calculate the conditional-composition data, where these ages are
conditional on a length bin.
}
}

\section{To Do}{
The following to do list is ordered by priority with the highest priority
items listed first.
\itemize{
\item Create a sensitivity that changes the arguments passed to
\code{\link[nwfscSurvey:SurveyLFs.fn]{nwfscSurvey::SurveyLFs.fn()}} of \code{sexRatioUnsexed = 0.50},
\code{maxSizeUnsexed = 32}, and \code{sexRatioStage = 2}. Then use these marginal
length compositions in the data file. Chantel suggested the following
figure as well

\if{html}{\out{<div class="sourceCode">}}\preformatted{nwfscSurvey::PlotFreqData.fn(
  dir =  figure_dir,
  dat = lf_sex_ratio,
  main = "NWFSC Groundfish Bottom Trawl Survey",
  ylim = c(0, max(len_bins)),
  yaxs = "i",
  ylab = "Length (cm)"
)
}\if{html}{\out{</div>}}
\item There is a mix of saving things as wcgbts and NWFSC.Combo. Increase
consistency.
\item \code{\link[=PlotBio.fn]{PlotBio.fn()}} and \code{\link[=PlotBioStrata.fn]{PlotBioStrata.fn()}} do not allow you to change
the name but this function could rename the output.
\item The current functionality attempts to do many things for each survey
included in \verb{data_survey_*} but this fails because the strata only works
for the WCGBTS. So, in the end many outputs are limited to just this most
recent survey. It would be nice to get the function to work for all
surveys.
}
}

